CLOUDP-135435: Follow ups (#1468)
